Move the plotly stuff to its own submodule
